chore: move classifier tests to integration tests
